Refactor getPlaylists into spotify.js
