+++ backend_src/http_api/spotify.js
@@ -35,7 +35,53 @@ const getPlaylistById = (playlistId: string) => {
   return data;
 }
 
+
+const getPlaylists = async ({ user, private }) => {
+  if (private == null) {
+    private = false;
+  }
+  
+  if (private) {
+    const { access_token } = await auth.refreshAccessToken(user.refreshToken);
+    await db.get().collection('users').updateOne({ _id: user._id }, { $set: { accessToken: access_token } });
+    try {
+      const { data } = await axios({
+        url: 'https://api.spotify.com/v1/me/playlists',
+        headers: {
+          'Authorization': 'Bearer ' + access_token,
+        }
+      })
+      return data.items.map(playlist => ({
+        id: playlist.id,
+        imageUrl: playlist.images[0].url,
+        name: playlist.name,
+      }));
+    } catch (err) {
+      throw err;
+    }
+  } else {
+    token = await auth.getToken();
+    try {
+      const { data } = await axios({
+        method: 'get',
+        url: `https://api.spotify.com/v1/users/${user._id}/playlists`,
+        headers: {
+          'Authorization': 'Bearer ' + token
+        },
+      })
+      return data.items.map(playlist => ({
+        id: playlist.id,
+        imageUrl: playlist.images[0].url,
+        name: playlist.name,
+      }));
+    } catch (err) {
+      throw err
+    }
+  }
+}
+
 module.exports = {
   getPlaylistById,
-  searchPlaylist
+  searchPlaylist,
+  getPlaylists
 }
